refactor(examples): refactor the examples of llm and add examlples of confts
