--- examples/llm/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) 2023-present, SUSTech-ML.
-# All rights reserved.
-#
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-#
-
