+++ torchcp/classification/trainer/__init__.py
@@ -10,4 +10,3 @@ from .ordinal_trainer import OrdinalTrainer
 from .ts_trainer import TSTrainer
 from .confts_trainer import ConfTSTrainer
 from .model import TemperatureScalingModel
-from .conflearn_trainer import ConfLearnTrainer
